requirements.txt
flask==3.1.2
gunicorn==21.2.0
flask-sqlalchemy
flask-migrate
flask-login
bcrypt
psycopg2-binary


